Refactor: split build-composition ability into collaborators [ED-24826]
Extract the 1510-line Build_Composition_Ability into single-responsibility
classes under Abilities\Build_Composition\ (Xml_Parser, Widget_Type_Resolver,
Subtree_Builder, Element_Config_Applier, Style_Applier, Persister). The
orchestrator now reads as a short pipeline (~330 lines).

Move the LLM prompt to static-resources/build-composition/prompt.md, loaded
via file_get_contents like Style_Best_Practices_Ability, so it can be reused
outside PHP later.

Remove leftover dev instrumentation: debug_log(), summarize_*_for_debug(),
error_log() per-step writes, .cursor/debug-518cfe.log appends, hardcoded
sessionId/hypothesisId markers, and the applied/skipped/prop_resolutions
bookkeeping arrays that only fed those logs.

Minor: add Prop_Canonicalizer::build_alias_map() and an optional $alias_map
param to resolve_canonical_key() so callers can precompute the map once per
schema. Split Llm_Prop_Value_Adjuster recursion into adjust/adjust_nested_value.

Architecture, execute() inputs/outputs, error codes, ability id, and the
constructor signature ( ?Document_Mutator ) are all preserved. Existing tests
untouched.

Ref: ED-24826
